Add PHPUnit test script
```
usage: ./bin/phpunit-docker.sh [wp-version]
```

Adds a new script that uses docker to create a temporary database
container, installs the WP Unit Test dependencies, and runs the tests
in a phpunit container.

We're also splitting the existing install-wp-tests.sh script into two parts
so the applicable parts are reusable for this. The part that sets up a
local database is unnecessary here since we're using Docker for that.

The existing script still references the new download-wp-tests.sh script
internally, so that should continue to function the same as before.

Fixes #1017